+++ packages/create-react-native-library/templates/cpp-library/android/cpp-adapter.cpp
@@ -3,6 +3,10 @@
 
 extern "C"
 JNIEXPORT jint JNICALL
-Java_com_<%- project.package -%>_<%- project.name -%>Module_nativeMultiply(JNIEnv *env, jclass type, jint a, jint b) {
-    return example::multiply(a, b);
+<% if (project.architecture == 'mixed') { -%>
+Java_com_<%- project.package_cpp -%>_<%- project.name -%>ModuleImpl_multiply(JNIEnv *env, jclass type, jdouble a, jdouble b) {
+<% } else { -%>
+Java_com_<%- project.package_cpp -%>_<%- project.name -%>Module_nativeMultiply(JNIEnv *env, jclass type, jdouble a, jdouble b) {
+<% } -%>
+    return <%- project.package_cpp -%>::multiply(a, b);
 }