+++ packages/create-react-native-library/templates/java-library-mixed/android/src/turbo/{%- project.name %}Module.java
@@ -15,7 +15,7 @@ public class <%- project.name -%>Module extends Native<%- project.name -%>Spec {
   @Override
   @NonNull
   public String getName() {
-    return <%- project.name -%>ModuleImpl.NAME;
+    return NAME;
   }
 
   // Example method
@@ -23,6 +23,6 @@ public class <%- project.name -%>Module extends Native<%- project.name -%>Spec {
   @Override
   @ReactMethod
   public void multiply(double a, double b, Promise promise) {
-    <%- project.name -%>ModuleImpl.multiply(a, b, promise);
+    promise.resolve(<%- project.name -%>ModuleImpl.multiply(a, b));
   }
 }