+++ packages/create-react-native-library/templates/kotlin-library-mixed/android/src/turbo/{%- project.name %}Module.kt
@@ -1,0 +1,23 @@
+package com.<%- project.package -%>;
+
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactMethod
+import com.facebook.react.bridge.Promise
+
+class <%- project.name -%>Module internal constructor(context: ReactApplicationContext?) :
+  Native<%- project.name -%>Spec(context) {
+  override fun getName(): String {
+    return NAME
+  }
+
+  // Example method
+  // See https://reactnative.dev/docs/native-modules-android
+  @ReactMethod
+  override fun multiply(a: Double, b: Double, promise: Promise) {
+    <%- project.name -%>ModuleImpl.multiply(a, b, promise)
+  }
+
+  companion object {
+    val NAME: String = <%- project.name -%>ModuleImpl.NAME
+  }
+}