+++ packages/react-native-builder-bob/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.20.0](https://github.com/callstack/react-native-builder-bob/compare/react-native-builder-bob@0.19.1...react-native-builder-bob@0.20.0) (2022-10-15)
+
+
+### Features
+
+* add a run command to bob ([16dace2](https://github.com/callstack/react-native-builder-bob/commit/16dace209521d25f64b8d40631bb74241444938c))
+
+
+
+
+
+## [0.19.1](https://github.com/callstack/react-native-builder-bob/compare/react-native-builder-bob@0.19.0...react-native-builder-bob@0.19.1) (2022-10-14)
+
+
+### Bug Fixes
+
+* restrict engine to latest Node LTS ([f569064](https://github.com/callstack/react-native-builder-bob/commit/f5690640cfca5b3b7c32b98c80442b08fda09150))
+
+
+
+
+
+# [0.19.0](https://github.com/callstack/react-native-builder-bob/compare/react-native-builder-bob@0.18.3...react-native-builder-bob@0.19.0) (2022-10-13)
+
+
+### Features
+
+* improve sourcemaps generation ([#280](https://github.com/callstack/react-native-builder-bob/issues/280)) ([cb404bf](https://github.com/callstack/react-native-builder-bob/commit/cb404bf6c784c59abebe5e0cb58c2d9a1b3b3020)), closes [#279](https://github.com/callstack/react-native-builder-bob/issues/279)
+
+
+
+
+
 ## [0.18.3](https://github.com/callstack/react-native-builder-bob/compare/react-native-builder-bob@0.18.2...react-native-builder-bob@0.18.3) (2022-06-21)
 
 **Note:** Version bump only for package react-native-builder-bob